chore: implementing health processor

+++ src/worker/src/main/kotlin/com/example/worker/app/models/Health.kt
@@ -2,5 +2,5 @@ package com.example.worker.app.models
 
 data class Health(
     val failing: Boolean,
-    val minResponseTime: Int
+    val minResponseTime: Long
 )
